backend/api/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
neo4j==5.15.0
openai==1.69.0
pydantic==2.6.1 